+++ modules/core/src/main/scala/doobie/package.scala
@@ -1,7 +1,8 @@
 package object doobie
   extends Aliases
-     with hi.Aliases
-     with free.Aliases {
+     with doobie.hi.Modules
+     with doobie.free.Modules
+     with doobie.free.Types {
 
   object implicits
     extends free.Instances
@@ -10,8 +11,9 @@ package object doobie
   @deprecated(message = "import doobie._, doobie.implicits._", since = "0.5.0")
   object imports
     extends Aliases
-      with hi.Aliases
-      with free.Aliases
+      with hi.Modules
+      with free.Modules
+      with free.Types
       with free.Instances
       with syntax.AllSyntax
