+++ modules/h2/src/main/scala/doobie/h2/Metas.scala
@@ -12,7 +12,7 @@ import scala.reflect.runtime.universe.TypeTag
 
 import cats.data.NonEmptyList.{ of => NonEmptyListOf }
 
-trait Metas {
+trait Instances {
 
   implicit val UuidType =
     Meta.advanced[UUID](
