+++ modules/h2/src/main/scala/doobie/h2/package.scala
@@ -7,12 +7,12 @@ package object h2 {
   type H2Transactor[M[_]] = Transactor.Aux[M, JdbcConnectionPool]
 
   object implicits
-    extends Metas
+    extends Instances
        with syntax.ToH2TransactorOps
 
   @deprecated(message = "import doobie.h2._, doobie.h2.implicits._", since = "0.5.0")
   object imports
-    extends Metas
+    extends Instances
        with syntax.ToH2TransactorOps {
     type H2Transactor[M[_]] = h2.H2Transactor[M]
   }
