+++ modules/postgres/src/main/scala/doobie/postgres/hi/aliases.scala
@@ -1,0 +1,8 @@
+package doobie.postgres.hi
+
+trait Modules {
+  val PHPC  = pgconnection
+  val PHC   = connection
+  val PHLO  = largeobject
+  val PHLOM = largeobjectmanager
+}
