+++ modules/postgres/src/main/scala/doobie/postgres/hi/connection.scala
@@ -5,7 +5,6 @@ import cats.data.Kleisli
 import org.postgresql.{ PGConnection, PGNotification }
 import doobie.imports._
 import doobie.postgres.free.KleisliInterpreter
-import doobie.postgres.imports._
 
 /** Module of safe `PGConnectionIO` operations lifted into `ConnectionIO`. */
 object connection {
