+++ modules/postgres/src/main/scala/doobie/postgres/hi/package.scala
@@ -1,0 +1,8 @@
+package doobie.postgres
+
+package object hi
+  extends Modules
+     with free.Modules
+     with free.Types {
+  object implicits extends free.Instances
+}
