+++ modules/postgres/src/test/scala/doobie/postgres/check.scala
@@ -1,8 +1,8 @@
 package doobie.postgres
 
 import cats.effect.IO
-import doobie.imports._
-import doobie.postgres.imports._
+import doobie._, doobie.implicits._
+import doobie.postgres._, doobie.postgres.implicits._
 import org.specs2.mutable.Specification
 
 object pgcheck extends Specification {
