+++ modules/postgres/src/test/scala/doobie/postgres/pgnotify.scala
@@ -2,8 +2,8 @@ package doobie.postgres
 
 import cats.effect.{ IO, Sync }
 import cats.implicits._
-import doobie.imports._
-import doobie.postgres.imports._
+import doobie._, doobie.implicits._
+import doobie.postgres._, doobie.postgres.implicits._
 import org.postgresql.PGNotification
 import org.specs2.mutable.Specification
