--- modules/scalatest/src/main/scala/doobie/scalatest/imports.scala
@@ -1,11 +1,0 @@
-package doobie.scalatest
-
-object imports {
-
-  type Checker[M[_]] = doobie.scalatest.Checker[M]
-  type IOChecker = doobie.scalatest.IOChecker
-
-  @deprecated("Use IOChecker.", "0.4.2")
-  type QueryChecker = IOChecker
-
-}
